Try not-too-hard to avoid aborting in benchmarks
So, if we pick a modulus of 0, of course uniform_1 croaks.  If our
generator really is acceptably random, that should be pretty rare.  We
add another call, which still doesn't guarantee we don't get two zeros
in a row, but if that happens, the PRNG is more likely to be broken or
your luck was really bad.  We could just bitwise-or in a 1, but I
prefer having this canary.

As evidence that I have tested this code too infrequently without
rdrand, it turns out that PCG32 initialization without rdrand has been
quite weak, all this time, and it's quite common to get 0 from
uniform32 here as the modulus, much more than it should have been.