Moderately improve PCG32 seeding without rdrand
Without rdrand, benchmarks were aborting because the first value
generated was often zero.  Having high-quality seeding has never been
a goal of this library, but this was pretty bad.  So we mix in an
initialization constant (from some other code I wrote using pcg32,
which presumably took it from upstream PCG examples or a similar
source).