Implement "rdtsc" for aarch64
We don't fill u, but then again, that's reasonably within the spirit
of this library.

Some testing revealed PMCCNTR_EL0 can't be read from userspace on
macOS, so we call the blessed function and hope that it doesn't turn
into an expensive syscall, though that shouldn't really matter for our
uses of rdtsc here.

Fixes #13.